Prevent process form shutting down when calling WebFrameProxy::prepareForProvisionalNavigationInProcess https://bugs.webkit.org/show_bug.cgi?id=265550 rdar://118471432
Reviewed by Pascoe.

Similar to code in WebPageProxy::triggerBrowsingContextGroupSwitchForNavigation
and WebPageProxy::continueNavigationInNewProcess, if we do anything asynchronous between
the point where we decide to use a process and the point where we actually start using it,
we need to make a shutdown preventing scope.  Otherwise, we hit the release assertion
in WebPageProxy::continueNavigationInNewProcess, which happened occasionally when loading
to and from complicated web pages with site isolation enabled.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::continueNavigationInNewProcess):

Canonical link: https://commits.webkit.org/271307@main