Correct <input type=checkbox switch> animation corner case and add tests https://bugs.webkit.org/show_bug.cgi?id=265356
Reviewed by Aditya Keerthi.

I've decided on the following things:

* When an animation is running and setChecked() is called with the same
  value the animation is running for, the animation continues to run.
  As such we do not modify the early return in setChecked() for when
  the state does not change unlike an earlier iteration of this change.
* When an animation is running and the checked or disabled state
  changes underneath, the animation is stopped.

Stopping the animation is abstracted into a method for clarity.

Test coverage is added for these scenarios and a couple more to reduce
the chances of regressions when refactoring in the future.

I also want to mention here that I strongly considered an alternative
approach whereby CheckboxInputType would be solely responsible for
calling performSwitchCheckedChangeAnimation() right after it calls
setChecked(), but setChecked() can invalidate or reverse an animation
and I could not find a good way to integrate that behavior.

Canonical link: https://commits.webkit.org/271327@main