Web Inspector: CSS Grid Page Overlay support for masonry layout https://bugs.webkit.org/show_bug.cgi?id=265381 rdar://118833618
Reviewed by Patrick Angle.

There are no actual rows or columns in the masonry axis of a masonry layout.
But we can borrow the concept to draw the two lines at the start and end of the masonry axis.

There are rows or columns in the grid axis of a masonry layout, just like in a regular grid layout.
Their lines will wrap the masonry layout in the other direction.

* Source/WebCore/inspector/InspectorOverlay.cpp:
(WebCore::InspectorOverlay::buildGridOverlay):

Canonical link: https://commits.webkit.org/271332@main