Element.cpp's findAttrNodeInList() does case-insensitive comparison incorrectly https://bugs.webkit.org/show_bug.cgi?id=265597
Reviewed by NOBODY (OOPS!).

If you call setAttributeNode() twice with an Attr nodes that have the same uppercase
attribute name (e.g. "FOOBAR"), in the context of an HTML document, it will fail to
find the the previous Attr node when the second call happens. As a result, setAttributeNode()
will return null twice instead of returning the existing Attr name with the same name the
second time around.

The bug was in findAttrNodeInList(). When trying to do a case insensitive check, it would
lower case the input localName and then compare it to the local names of existing attributes,
without lower casing those. As a result, we would compare "foobar" to "FOOBAR" and fail to
match.

Thanks to Claudio Saavedra from Igalia for finding the bug and doing much of the investigation
on this.

* LayoutTests/fast/dom/Element/setAttributeNode-same-attribute-name-uppercase-expected.txt: Added.
* LayoutTests/fast/dom/Element/setAttributeNode-same-attribute-name-uppercase.html: Added.
* Source/WebCore/dom/Element.cpp:
(WebCore::findAttrNodeInList):