[UIAsyncTextInput] Release assertion when dismissing the keyboard or changing focused element https://bugs.webkit.org/show_bug.cgi?id=265736
Reviewed by Tim Horton.

Simply delete this codepath, instead of release asserting. The only call site of this method is from
`UIDictationController` in UIKit, which is gated on a `-respondsToSelector:` check.

Since this isn't needed for binary compatibility with any other internal apps, and the resulting
behavior in `-[UIDictationController logCorrectionStatisticsForDelegate:reason:]` is unchanged, we
should just delete this code altogether.

* Source/WebKit/Platform/spi/ios/UIKitSPI.h:
* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView metadataDictionariesForDictationResults]): Deleted.

Canonical link: https://commits.webkit.org/271442@main