[Cocoa] Fullscreen on vice.com: CC font size is extremely large in portrait view https://bugs.webkit.org/show_bug.cgi?id=265817 rdar://118030141
Reviewed by Simon Fraser.

When a landscape video is presented in portrait mode (with large letterboxes), this causes the font size of the
associated text tracks to be absurdly large. Filed a spec issue on WebVTT (https://github.com/w3c/webvtt/issues/522)
to change the default font size be `5vmin` rather than `5vh`, but proceeding with changing WebKit's default
font size to be `5cqh` in the meantime.

* Source/WebCore/Modules/modern-media-controls/controls/text-tracks.css:
(::-webkit-media-text-track-region):

Canonical link: https://commits.webkit.org/271536@main