REGRESSION: Two editing/caret/color-span-inside-editable tests are constant image failures https://bugs.webkit.org/show_bug.cgi?id=265871 rdar://113875623
Reviewed by Brent Fulgham.

There were 2 caret-related changes:
A) Caret shape changed between Ventura and Sonoma (from squared to rounded).
B) Caret color behavior changed. The caret color should now update in content
editable elements if a foreground color is specified. This change was introduced by https://commits.webkit.org/266070@main.

Because of (A) we need to rebaseline these tests for the expected values
to match the shape of the updated caret. This change is only valid from Sonoma,
so we are adding -expected values for Ventura and Monterey, that expect
the caret to still be squared.

Because of (B) we need to rebaseline these tests for the expected values
to match the color of the caret in content editable elements. We also need
to update the description of the test themselves to match the new behavior, since
before the WebKit change on (B) caret would preserve its color even if set
by author.

* LayoutTests/editing/caret/color-span-inside-editable-background-expected.html:
* LayoutTests/editing/caret/color-span-inside-editable-background.html:
* LayoutTests/editing/caret/color-span-inside-editable-expected.html:
* LayoutTests/editing/caret/color-span-inside-editable.html:
* LayoutTests/platform/gtk/editing/caret/color-span-inside-editable-background-expected.html: Added.
* LayoutTests/platform/gtk/editing/caret/color-span-inside-editable-expected.html: Added.
* LayoutTests/platform/mac-monterey/editing/caret/color-span-inside-editable-background-expected.html: Added.
* LayoutTests/platform/mac-monterey/editing/caret/color-span-inside-editable-expected.html: Added.
* LayoutTests/platform/mac-ventura/editing/caret/color-span-inside-editable-background-expected.html: Added.
* LayoutTests/platform/mac-ventura/editing/caret/color-span-inside-editable-expected.html: Added.
* LayoutTests/platform/mac-wk2/editing/caret/caret-color-expected.png: Removed.
* LayoutTests/platform/mac-wk2/editing/caret/color-span-inside-editable-background-expected.html: Removed.
* LayoutTests/platform/mac-wk2/editing/caret/color-span-inside-editable-expected.html: Removed.
* LayoutTests/platform/mac/TestExpectations:

Canonical link: https://commits.webkit.org/271563@main