AX: The target element of an aria-labeledby relation may not have an associated AX object created. https://bugs.webkit.org/show_bug.cgi?id=265625 <rdar://problem/119013970>
Reviewed by Tyler Wilcock.

In the following example:

<button aria-labeledby="label"/> (1)
<span hidden id="label">z</span> (2)

We don't create an AX object that represents the element (2) since it is hidden. Other examples can be found in

    text-alternative-calculation-from-unrendered-table.html
and
    aria-labeled-with-hidden-node.html

ARIA specifies that we still have to expose (2) as the textual representation of (1). This creates a problem when attempting to fully implement LabeledBy relationships based on AX objects. This patch addresses the problem by unconditionally creating AX objects for elements that are targets of relationships.

* LayoutTests/accessibility/text-alternative-calculation-from-unrendered-table-expected.txt:
* LayoutTests/accessibility/text-alternative-calculation-from-unrendered-table.html: Modernized the code of this test.
* LayoutTests/platform/mac/accessibility/text-alternative-calculation-from-unrendered-table-expected.txt:
* LayoutTests/platform/mac-wk1/TestExpectations:
* Source/WebCore/accessibility/AXObjectCache.cpp:
(WebCore::AXObjectCache::getOrCreate):
(WebCore::AXObjectCache::addRelation):
(WebCore::AXObjectCache::isDescendantOfRelationTarget):
* Source/WebCore/accessibility/AXObjectCache.h:
(WebCore::AXObjectCache::getOrCreate):

Canonical link: https://commits.webkit.org/271566@main