[Flatpak SDK] Fix run-minibrowser for python 3.12 https://bugs.webkit.org/show_bug.cgi?id=265931
Reviewed by NOBODY (OOPS!).

When the host's Python version is 3.12 or higher, Flatpak SDK's run-minibrowser fails
due to conflicting setuptools versions (68.1.2 for Python >= 3.12 and 56.0.0 for
Python >= 3.0). To resolve, ensure matching setuptools versions between host and
Flatpak Pythons. The current SDK's Python version is 3.11.5, limited to setuptools-56.0.0
due to a beautifulsoup4 dependency since commit 267461@main. A similar adjustment
is made for Python 3.12.

* Tools/Scripts/libraries/webkitcorepy/webkitcorepy/__init__.py: