[offlineasm] Incremental build failure when generator is interrupted https://bugs.webkit.org/show_bug.cgi?id=265729 rdar://119079705
Reviewed by Alexey Proskuryakov.

offlineasm writes headers non-atomically, so if the script is
interrupted due to a build failure in a parallel task, it leaves a
partially-written header. The build system can't tell that this header
is incomplete, so its presence breaks subsequent rebuilds.

Fix by writing to a temporary file in the same directory (or
TARGET_TEMP_DIR in Xcode), and `rename`-ing once generation is complete.

* Source/JavaScriptCore/offlineasm/generate_offset_extractor.rb:
* Source/JavaScriptCore/offlineasm/generate_settings_extractor.rb:

Canonical link: https://commits.webkit.org/271640@main