[UnifiedPDF] PDFs load blank/white on iOS with UnifiedPDF enabled https://bugs.webkit.org/show_bug.cgi?id=265966 rdar://118522690
Reviewed by Tim Horton.

The `PDFPluginEnabled` preference defaults to OFF for IOS_FAMILY
targets, so even when the `UnifiedPDFEnabled` preference is set, on iOS we
always report `shouldUsePDFPlugin==false`, which means a plugin is never
created and we never see a PDF load.

This patch fixes this bug by consulting whether or not the
`UnifiedPDFEnabled` has been set. If it has, then we should use the PDF
plugin.

* Source/WebKit/WebProcess/WebPage/Cocoa/WebPageCocoa.mm:
(WebKit::WebPage::shouldUsePDFPlugin const):

Canonical link: https://commits.webkit.org/271660@main