[web-animations] compositing/layer-creation/scale-rotation-animation-overlap.html is a timeout on Sonoma https://bugs.webkit.org/show_bug.cgi?id=266171
Reviewed by Simon Fraser.

We used to create a 60x60 dot matrix and this would cost a lot of the test operations
to be insanely slow in Debug builds. Using a 30x30 matrix (4 times fewer elements) with
wider spacing covers the same area but dramatically shortens the time it takes to run the
test, going from about 30 seconds on my development laptop to just about 2 seconds and thus
comfortably avoiding a timeout.

* LayoutTests/compositing/layer-creation/scale-rotation-animation-overlap-expected.txt:
* LayoutTests/compositing/layer-creation/scale-rotation-animation-overlap.html:
* LayoutTests/platform/ios/compositing/layer-creation/scale-rotation-animation-overlap-expected.txt:

Canonical link: https://commits.webkit.org/271845@main